Bug fix in update_head function :
update_head function is supposed to update the table with
latest applied head. But it was adding new rows instead.
Reason is I was trying to update the primary key which was a
mistale. Now A different primary key has been added in the
table structure and it works fine.

A few indentation related changes have been added too.